USB库启用后CPU占有率99%问题莫名其妙好了
先备份再查问题 30ms读取一张光谱 估计大多数时间时间是在拷贝数据

+++ USB_Host_Library/Core/Src/usbh_ioreq.c
@@ -1,27 +1,47 @@
-/** 
+/**
   ******************************************************************************
-  * @file    usbh_ioreq.c 
+  * @file    usbh_ioreq.c
   * @author  MCD Application Team
   * @brief   This file handles the issuing of the USB transactions
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; COPYRIGHT 2015 STMicroelectronics</center></h2>
+  * <h2><center>&copy; Copyright (c) 2017 STMicroelectronics International N.V.
+  * All rights reserved.</center></h2>
   *
-  * Licensed under MCD-ST Liberty SW License Agreement V2, (the "License");
-  * You may not use this file except in compliance with the License.
-  * You may obtain a copy of the License at:
+  * Redistribution and use in source and binary forms, with or without
+  * modification, are permitted, provided that the following conditions are met:
   *
-  *        http://www.st.com/software_license_agreement_liberty_v2
+  * 1. Redistribution of source code must retain the above copyright notice,
+  *    this list of conditions and the following disclaimer.
+  * 2. Redistributions in binary form must reproduce the above copyright notice,
+  *    this list of conditions and the following disclaimer in the documentation
+  *    and/or other materials provided with the distribution.
+  * 3. Neither the name of STMicroelectronics nor the names of other
+  *    contributors to this software may be used to endorse or promote products
+  *    derived from this software without specific written permission.
+  * 4. This software, including modifications and/or derivative works of this
+  *    software, must execute solely and exclusively on microcontroller or
+  *    microprocessor devices manufactured by or for STMicroelectronics.
+  * 5. Redistribution and use of this software other than as permitted under
+  *    this license is void and will automatically terminate your rights under
+  *    this license.
   *
-  * Unless required by applicable law or agreed to in writing, software 
-  * distributed under the License is distributed on an "AS IS" BASIS, 
-  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  * See the License for the specific language governing permissions and
-  * limitations under the License.
+  * THIS SOFTWARE IS PROVIDED BY STMICROELECTRONICS AND CONTRIBUTORS "AS IS"
+  * AND ANY EXPRESS, IMPLIED OR STATUTORY WARRANTIES, INCLUDING, BUT NOT
+  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+  * PARTICULAR PURPOSE AND NON-INFRINGEMENT OF THIRD PARTY INTELLECTUAL PROPERTY
+  * RIGHTS ARE DISCLAIMED TO THE FULLEST EXTENT PERMITTED BY LAW. IN NO EVENT
+  * SHALL STMICROELECTRONICS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+  * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+  * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
+  * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+  * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
   *
   ******************************************************************************
-  */ 
+  */
 /* Includes ------------------------------------------------------------------*/
 
 #include "usbh_ioreq.h"
@@ -33,8 +53,8 @@
 /** @addtogroup USBH_LIB_CORE
 * @{
 */
-  
-/** @defgroup USBH_IOREQ 
+
+/** @defgroup USBH_IOREQ
   * @brief This file handles the standard protocol processing (USB v2.0)
   * @{
   */
@@ -42,48 +62,48 @@
 
 /** @defgroup USBH_IOREQ_Private_Defines
   * @{
-  */ 
+  */
 
 /**
   * @}
-  */ 
- 
+  */
+
 
 /** @defgroup USBH_IOREQ_Private_TypesDefinitions
   * @{
-  */ 
+  */
 /**
   * @}
-  */ 
+  */
 
 
 
 /** @defgroup USBH_IOREQ_Private_Macros
   * @{
-  */ 
+  */
 /**
   * @}
-  */ 
+  */
 
 
 /** @defgroup USBH_IOREQ_Private_Variables
   * @{
-  */ 
+  */
 /**
   * @}
-  */ 
+  */
 /** @defgroup USBH_IOREQ_Private_FunctionPrototypes
   * @{
-  */ 
+  */
 
 /**
   * @}
-  */ 
+  */
 
 
 /** @defgroup USBH_IOREQ_Private_Functions
   * @{
-  */ 
+  */
 
 
 
@@ -95,20 +115,20 @@
   * @param  pipe_num: Pipe Number
   * @retval USBH Status
   */
-USBH_StatusTypeDef USBH_CtlSendSetup (USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_CtlSendSetup (USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint8_t pipe_num)
 {
 
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT  */
+                          0U,                    /* Direction : OUT  */
                           USBH_EP_CONTROL,      /* EP type          */
                           USBH_PID_SETUP,       /* Type setup       */
                           buff,                 /* data buffer      */
-                          USBH_SETUP_PKT_SIZE,  /* data length      */ 
-                          0);
-  return USBH_OK;  
+                          USBH_SETUP_PKT_SIZE,  /* data length      */
+                          0U);
+  return USBH_OK;
 }
 
 
@@ -121,26 +141,26 @@ USBH_StatusTypeDef USBH_CtlSendSetup (USBH_HandleTypeDef *phost,
   * @param  pipe_num: Pipe Number
   * @retval USBH Status
   */
-USBH_StatusTypeDef USBH_CtlSendData (USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_CtlSendData (USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint16_t length,
                                 uint8_t pipe_num,
                                 uint8_t do_ping )
 {
   if(phost->device.speed != USBH_SPEED_HIGH)
   {
-    do_ping = 0;
+    do_ping = 0U;
   }
-  
+
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT  */
-                          USBH_EP_CONTROL,      /* EP type          */                          
+                          0U,                   /* Direction : OUT  */
+                          USBH_EP_CONTROL,      /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */ 
+                          length,               /* data length      */
                           do_ping);             /* do ping (HS Only)*/
-  
+
   return USBH_OK;
 }
 
@@ -152,23 +172,23 @@ USBH_StatusTypeDef USBH_CtlSendData (USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the response needs to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_CtlReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t* buff, 
+USBH_StatusTypeDef USBH_CtlReceiveData(USBH_HandleTypeDef *phost,
+                                uint8_t* buff,
                                 uint16_t length,
                                 uint8_t pipe_num)
 {
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_CONTROL,      /* EP type          */                          
+                          1U,                    /* Direction : IN   */
+                          USBH_EP_CONTROL,      /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */ 
-                          0);
+                          length,               /* data length      */
+                          0U);
   return USBH_OK;
-  
+
 }
 
 
@@ -181,24 +201,24 @@ USBH_StatusTypeDef USBH_CtlReceiveData(USBH_HandleTypeDef *phost,
   * @param  pipe_num: Pipe Number
   * @retval USBH Status
   */
-USBH_StatusTypeDef USBH_BulkSendData (USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_BulkSendData (USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint16_t length,
                                 uint8_t pipe_num,
-                                uint8_t do_ping )
-{ 
+                                uint8_t do_ping)
+{
   if(phost->device.speed != USBH_SPEED_HIGH)
   {
-    do_ping = 0;
+    do_ping = 0U;
   }
-  
+
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : IN   */
-                          USBH_EP_BULK,         /* EP type          */                          
+                          0U,                    /* Direction : IN   */
+                          USBH_EP_BULK,         /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */  
+                          length,               /* data length      */
                           do_ping);             /* do ping (HS Only)*/
   return USBH_OK;
 }
@@ -211,21 +231,21 @@ USBH_StatusTypeDef USBH_BulkSendData (USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the received data packet to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_BulkReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_BulkReceiveData(USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint16_t length,
                                 uint8_t pipe_num)
 {
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_BULK,         /* EP type          */                          
+                          1U,                    /* Direction : IN   */
+                          USBH_EP_BULK,         /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          0);
+                          length,               /* data length      */
+                          0U);
   return USBH_OK;
 }
 
@@ -237,22 +257,22 @@ USBH_StatusTypeDef USBH_BulkReceiveData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the response needs to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint8_t length,
                                 uint8_t pipe_num)
 {
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_INTERRUPT,    /* EP type          */                          
+                          1U,                   /* Direction : IN   */
+                          USBH_EP_INTERRUPT,    /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          0); 
-  
+                          (uint16_t)length,     /* data length      */
+                          0U);
+
   return USBH_OK;
 }
 
@@ -263,22 +283,22 @@ USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer from where the data needs to be copied
   * @param  length: Length of the data to be sent
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint8_t length,
                                 uint8_t pipe_num)
 {
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT   */
-                          USBH_EP_INTERRUPT,    /* EP type          */                          
+                          0U,                   /* Direction : OUT   */
+                          USBH_EP_INTERRUPT,    /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          0);  
-  
+                          (uint16_t)length,     /* data length      */
+                          0U);
+
   return USBH_OK;
 }
 
@@ -289,23 +309,23 @@ USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the response needs to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_IsocReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_IsocReceiveData(USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint32_t length,
                                 uint8_t pipe_num)
-{    
+{
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_ISO,          /* EP type          */                          
+                          1U,                   /* Direction : IN   */
+                          USBH_EP_ISO,          /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */
-                          0);
+                          (uint16_t)length,     /* data length      */
+                          0U);
 
-  
+
   return USBH_OK;
 }
 
@@ -316,31 +336,27 @@ USBH_StatusTypeDef USBH_IsocReceiveData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer from where the data needs to be copied
   * @param  length: Length of the data to be sent
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_IsocSendData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
+USBH_StatusTypeDef USBH_IsocSendData(USBH_HandleTypeDef *phost,
+                                uint8_t *buff,
                                 uint32_t length,
                                 uint8_t pipe_num)
 {
   USBH_LL_SubmitURB (phost,                     /* Driver handle    */
                           pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT   */
-                          USBH_EP_ISO,          /* EP type          */                          
+                          0U,                   /* Direction : OUT   */
+                          USBH_EP_ISO,          /* EP type          */
                           USBH_PID_DATA,        /* Type Data        */
                           buff,                 /* data buffer      */
-                          length,               /* data length      */ 
-                          0);
-  
+                          (uint16_t)length,     /* data length      */
+                          0U);
+
   return USBH_OK;
 }
 /**
 * @}
-*/ 
-
-/**
-* @}
-*/ 
+*/
 
 /**
 * @}
@@ -348,7 +364,11 @@ USBH_StatusTypeDef USBH_IsocSendData(USBH_HandleTypeDef *phost,
 
 /**
 * @}
-*/ 
+*/
+
+/**
+* @}
+*/
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
